refactor(BA-2619): Ensure model fields are set pre-validation
Modifying Pydantic model objects after validation (which is the case
when validators are set to "after" mode) is an anti-pattern, as Pydantic
expects the model to be immutable once values are set.

This change modifies the validators so that it runs in "before" mode,
meaning that any "computed" or "overridden" data will be set before any
validation is done.