fix(BA-2619): Add hierarchy to config context
The ConfigValidationContext object was mistakenly placed in
logging/config directory, when the config validation context should live
in a common directory independent from logging configs. Also, the config
validation context should not contain agent server-specific parameters,
and logging config should be independent from agent config context.

This change introduces a BaseConfigValidationContext which contains
common parameters that could be used for all script configs, and
AgentConfigValidationContext which is a specific instantiation of the
base class specifically for server.py of agent.