feat(BA-2606): Add subagent layer to agents
This change introduces a new abstraction layer named 'subagents', which
is used to emulate multiple agent backend instances within the same
agent deployment.

Users are now able to specify sub_agents in the unified configuration,
which enables the user to list the configurations of subagent instances
intended to be spawned. The agent server automatically handles creating
the agent instances according to the configuration and routing RPC calls
appropriately.

This current change does not yet fully handle routing RPC calls to the
correct subagent. If kernel ID, image name etc. are included as part of
the arguments of the RPC call, the requests will be directed to the
correct subagent. Otherwise it currently routes the requests to the
default agent, which in this change is defined to be the first subagent
defined in the configuration.

Including subagent ID in the RPC request is future work.