fix(BA-2606): Add support for TOML table arrays in sample generator
This change adds support for generating TOML array of tables in the
sample config generator, which is required for generating subagents
configuration.

This change also fixes some subtle bugs with not including the
descriptions of certain optional fields.

Note that the sample_generator.py file has been vibe coded with Claude
Code, so maybe there are some subtle bugs with the generation. Care has
been taken to ensure that the code does generate correct config file.