fix(BA-2606): Fix config type passed into agent constructor
This change fixes an uncaught error with agent creation, where the type
of the config passed into the constructor was separated
AgentGlobalConfig and AgentSpecificConfig, rather than
AgentUnifiedConfig. This was a remnant of an older version of this
change, where the constructor type of the AbstractAgent was modified
which broke some implicit contracts of subclasses and how they used
config objects (especially with dumping the config in pickle).