feat(BA-2750): Add support for array in config generator
This change adds array of tables syntax [[table]] of TOML in the sample
config generator. As the downstream change of adding multiple agents in
agent runtime server will need the config to be able to express array of
tables, having the sample generator be able to handle this would be
beneficial.