fix(BA-2752): Move incorrectly placed agent config
This change moves any incorrectly placed agent config into the correct
location. Correct here means that the shared configs are supposed to be
shared, and per-agent unique fields are placed in the overridables
classes.

This change also fixes a pre-existing issue with how scaling group is
updated, where the local config stored in the agent server is updated,
but the change does not get propagated to the individual AbstractAgent.
This issue already existed, but prior to this change this PR also did
not address it. This change fixes it.