fix(BA-2750): Ensure list configs are not sorted
As list fields e.g. allocation_order need to preserve their order, as
ordering is important for them, we should not sort them. Only set
values, which by nature of the type means that the order at which the
examples are displayed don't matter, should be sorted to prevent
non-determinism.