fix(BA-2753): Remove segregation of path per agent
Certain paths like IPC, var, etc... are not actually meant to be
separated out to be overridable per agent, but rather shared by all
agents in agent runtime. This change reverts any change that assumed
that the paths will be separated, and it also adds fixes where the files
written in these directories were not unique per agent.