refactor(BA-3028): Move kernel registry ownership to AgentRuntime
This change moves where the agent's kernel registry is stored from
within the agent to outside the agent at a global level in AgentRuntime.
This is to make future change of pulling out Metadata server outside of
DockerAgent easier, which needs to take in a global view of all kernels
across all agents.