fix(BA-2851): Exclude reserved resources from total only when needed
This change fixes a bug with resource splitting, where reserved
resources were accidentally being included in the total allocated for
each agent. This is because the way total slots are handled was
malformed, where the calculation of reserved resources from the
perspective of a single agent was being done without taking account of
server reserved resources properly. This change fixes this issue by
inverting the condition, where reserved resources are deducted only in
places where it is needed.