refactor(BA-3072): Enforce interface consistency for all agents
This change makes refactoring changes to ensure that all @abstractmethod
declarations in AbstractAgent and AbstractKernelCreationContext are all
marked with @override in the concrete implementation Agent classes. This
change was made in part of an investigation into what functionalities in
DockerAgent are missing in KubernetesAgent.

While this change is mostly adding @override decorators for functions,
it does address a couple of genuine discrepancies in the implementation
of DummyAgent and KubernetesAgent that are inconsistent with
AbstractAgent or DockerAgent.
* Spurious declaration of function extract_command(), which likely is an
  accidental residue change that added extract_image_command() method.
* KubernetesAgent was missing methods push_image() and
  extract_image_command().