refactor(BA-3072): Move uid/gid overrides to AbstractAgent
The fields overriding_uid/gid and supplementary_gids are currently
defined in DockerAgent. However, there is no reason that this should not
exist in AbstractAgent. The reason for the current implemtation was
because the behaviour could not be guaranteed on KubernetesAgent, which
is why this was moved to DockerAgent.

As there is an ongoing change of reimplementing KubernetesAgent to have
full feature-parity with existing DockerAgent, these functions should be
moved to AbstractAgent.