fix(BA-3088): Ensure no default ID set until instance ID fetched
There was a bug in the Agent config parsing, where the default agent ID
value was being set before it could be overridden with the instance ID
of the nodes. This change fixes this issue by making sure that the agent
ID on the config side remains optional, and we set defaults only after
the server had the chance of setting the agent ID to the ID of the node
instance.